+++ backend/src/users/users.service.ts
@@ -1,15 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from 'src/entities/users/user.entity';
 import { Repository } from 'typeorm';
 import { genSaltSync, hashSync} from "bcrypt";
 import { createUser } from './dto/create-user.dto';
+import { CaslAbilityFactory } from 'src/casl/casl-ability.factory';
+import { Action } from 'src/policies/action.enum';
 
 @Injectable()
 export class UsersService {
     constructor(
         @InjectRepository(User)
-        private readonly repo: Repository<User>
+        private readonly repo: Repository<User>,
+        private readonly caslAbilityFactory: CaslAbilityFactory
     ){}
     
     findById(id: number): Promise<User | undefined> {
@@ -26,16 +29,26 @@ export class UsersService {
         return this.repo.save(user);
     }
 
-    async update(id: number,dto: Partial<Omit<User, 'id' | 'password'>>): Promise<User>{
-        const user = { ...(await this.repo.findOne(id)), ...dto };
+    async update(user: User,id: number,dto: Partial<Omit<User, 'id' | 'password'>>): Promise<User>{
+        const updatedUser = { ...(await this.repo.findOne(id)), ...dto };
+        const ability = this.caslAbilityFactory.createForUser(user);
+        if(updatedUser == undefined) throw new NotFoundException();
+        if(!ability.can(Action.Update,updatedUser) && !(updatedUser.id == user.id)) throw new UnauthorizedException();
         if (dto.hashedPassword) {
-            user.hashedPassword = this.hash(dto.hashedPassword);
+            updatedUser.hashedPassword = this.hash(dto.hashedPassword);
         }
-        return this.repo.save(user);
+        return this.repo.save(updatedUser);
     }
 
-    async delete(id: number): Promise<User>{
-        const user = await this.repo.findOne(id);
+    async delete(user: User,id: number): Promise<User>{
+        const ability = this.caslAbilityFactory.createForUser(user);
+        const deleteUser = await this.repo.findOne(id);
+        if(deleteUser == undefined) throw new NotFoundException();
+        if(!ability.can(Action.Update,deleteUser)) throw new UnauthorizedException();
+
+        deleteUser.jobAnnouncements = [];
+        this.repo.save(deleteUser);
+        
         await this.repo.remove(user);
         return user;
     }
